Refactor - move aws common code to package
Common code makes sense to be in it's own package and not `utils`.

Also made common AWS vars live inside that package instead of the
commands.
